Don't require em-http and yajl if streaming is not needed.
